requirements_tests.txt
# -*- coding: utf-8 -*-
# ----------------------------------------------------------------------
# Wordless: CI - Requirements File - Tests
# Copyright (C) 2018-2023  Ye Lei (叶磊)
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <http://www.gnu.org/licenses/>.
# ----------------------------------------------------------------------

# NLP
botok == 0.8.10
charset-normalizer == 3.0.1
jieba == 0.42.1
lingua-language-detector == 1.1.3
nltk == 3.7
pyphen == 0.13.1
pythainlp == 3.1.1
sacremoses == 0.0.53
simplemma == 0.9.0
textblob == 0.17.1
underthesea == 1.3.5

## pymorphy2
pymorphy2 == 0.9.1
pymorphy2-dicts-ru == 2.4.417127.4579844
pymorphy2-dicts-uk == 2.4.1.1.1460299261

## spaCy
spacy == 3.4.3
spacy-lookups-data == 1.0.3
spacy-pkuseg == 0.0.32

## SudachiPy
sudachipy == 0.6.6
sudachidict_core == 20221021

# Miscellaneous
beautifulsoup4
lxml
matplotlib
networkx
numpy
opencc-python-reimplemented
openpyxl
pypdf2
pyqt5
pytest
python-docx
requests
scipy
wordcloud
